Refactor #520 - remove EffectOwner, ReduceOwner
