+++ statekit/core/src/main/kotlin/com/nlab/statekit/internal/MergeHelpers.kt
@@ -19,14 +19,9 @@ package com.nlab.statekit.internal
 /**
  * @author Thalys
  */
-
-// Test OK @see {com.nlab.statekit.internal.MergeHelpersKtTest}
-// TODO remove Generated annotation after deploy below issue
-// https://github.com/jacoco/jacoco/pull/1670
-@ExcludeFromGeneratedTestReport
 internal inline fun <T : Any> List<T>.merge(
     onMerge: (head: T, tails: List<T>) -> T,
-): T? =  when (size) {
+): T? = when (size) {
     0 -> null
     1 -> first()
     else -> onMerge(first(), drop(1))
